+++ WizardTemplate.java
@@ -1,9 +1,9 @@
 public class WizardTemplate extends Template{
     private double fireDelay;
-    private int clipSize;
+    private double clipSize;
     private double reloadTime;
     private double projectileSpeed;
-    private int projectileSize;
+    private double projectileSize;
     private double spread;
     public WizardTemplate(double moveSpeed,double maxHP,double armor,double regenRate,double regenDelay,double fireDelay,int clipSize,double reloadTime,double projectileSpeed,int projectileSize,double spread){
         super(moveSpeed,maxHP,armor,regenRate,regenDelay);
@@ -15,7 +15,7 @@ public class WizardTemplate extends Template{
         this.spread = spread;
     }
 
-    public void updateWeapon(double fireDelay,int clipSize,double reloadTime,double projectileSpeed,int projectileSize,double spread){
+    public void updateWeapon(double fireDelay,double clipSize,double reloadTime,double projectileSpeed,double projectileSize,double spread){
         this.fireDelay = fireDelay;
         this.clipSize = clipSize;
         this.reloadTime = reloadTime;
@@ -27,7 +27,7 @@ public class WizardTemplate extends Template{
     public double getFireDelay(){
         return this.fireDelay;
     }
-    public int getClipSize(){
+    public double getClipSize(){
         return this.clipSize;
     }
     public double getReloadTime(){
@@ -36,7 +36,7 @@ public class WizardTemplate extends Template{
     public double getProjectileSpeed(){
         return this.projectileSpeed;
     }
-    public int getProjectileSize(){
+    public double getProjectileSize(){
         return this.projectileSize;
     }
     public double getSpread(){
